\docType{package}
\name{wnwr}
\alias{wnwr}
\alias{wnwr-package}
\title{wnwr provides ...}
\description{
wnwr provides ...
}
\section{Package options}{
  It uses the following \code{options} to configure
  behaviour: \itemize{ \item
  \code{wnwr.supported.search.types}: vector of supported
  search types \item \code{wnwr.supported.search.opts}:
  vector of supported search options \item
  \code{wnwr.supported.synset.types}: vector of supported
  synset types \item \code{wnwr.wn.command}: wordnet shell
  command }
}

